Refactor fang state task and task (#126)
* refactor FangTaskState and Task

* update diesel 2.1 and diesel-derive-enum 2.1

* fix tests

* fix spelling and move use Uuid to top
