+++ Cargo.toml
@@ -34,12 +34,12 @@ typetag = "0.2"
 uuid = { version = "1.1", features = ["v4"] }
 
 [dependencies.diesel]
-version = "2.0"
+version = "2.1"
 features = ["postgres", "serde_json", "chrono", "uuid", "r2d2"]
 optional = true
 
 [dependencies.diesel-derive-enum]
-version = "2.0.1"
+version = "2.1"
 features = ["postgres"]
 optional = true
