+++ src/blocking.rs
@@ -1,12 +1,10 @@
 mod error;
-pub mod fang_task_state;
 pub mod queue;
 pub mod runnable;
 pub mod schema;
 pub mod worker;
 pub mod worker_pool;
 
-pub use fang_task_state::FangTaskState;
 pub use queue::*;
 pub use runnable::Runnable;
 pub use schema::*;
